// mc_frontend/openapi/model/annisResponse.ts
/**
 * Machina Callida Backend REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FrequencyItem } from './frequencyItem';
import { TextComplexity } from './textComplexity';
import { GraphData } from './graphData';
import { Solution } from './solution';


/**
 * A response with graph data from ANNIS, possibly with additional data for exercises.
 */
export interface AnnisResponse { 
    /**
     * Unique identifier (UUID) for the exercise.
     */
    exercise_id?: string;
    /**
     * Type of exercise, concerning interaction and layout.
     */
    exercise_type?: string;
    /**
     * List of items with frequency data for linguistic phenomena.
     */
    frequency_analysis?: Array<FrequencyItem>;
    graph_data?: GraphData;
    /**
     * Correct solutions for this exercise.
     */
    solutions?: Array<Solution>;
    text_complexity?: TextComplexity;
    /**
     * URI for accessing the exercise in this API.
     */
    uri?: string;
}