// mc_frontend/openapi/model/solution.ts
/**
 * Machina Callida Backend REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SolutionElement } from './solutionElement';


/**
 * Correct solution for an exercise.
 */
export interface Solution { 
    target?: SolutionElement;
    value?: SolutionElement;
}